Finished Address Conversion Logic
Completed logic for address conversions
Changed when parsing variables, to store results as integers rather than
strings
Error when using long hand form of parameters (for example, "-b 128"
will work fine, but when using "--partition-start=128" the value is not
stored in the variable, and thus the function does not display
correctly.
Added commented out test to display Address Command array to debug this
later.